move factory from data_request to general pymorize directory
