+++ src/pymorize/data_request/collection.py
@@ -7,8 +7,9 @@ from typing import Dict
 
 import deprecation
 
+from pymorize.factory import MetaFactory
+
 from ..utils import download_json_tables_from_url, list_files_in_directory
-from .factory import MetaFactory
 from .table import CMIP6DataRequestTable, CMIP7DataRequestTable, DataRequestTable
 from .variable import CMIP7DataRequestVariable
